+++ src/App.css
@@ -4,52 +4,6 @@
   box-sizing: border-box;
 }
 
-body {
+/* body {
   background-image: url(https://www.teahub.io/photos/full/51-518770_maths-pics.jpg);
-}
-
-.calculator-container {
-  border-radius: 5px;
-  width: 320px;
-  height: 312px;
-  margin: 100px auto;
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-}
-
-.screen {
-  border-radius: 5px 5px 0 0;
-  padding-top: 20px;
-  padding-right: 3px;
-  width: 100%;
-  background-color: grey;
-  color: aliceblue;
-  text-align: right;
-  grid-column: 1/5;
-  font-size: 18px;
-}
-
-.width {
-  width: 100%;
-  height: 100%;
-  font-size: 17px;
-  border: none;
-}
-
-.width:hover {
-  background-color: aquamarine;
-  cursor: pointer;
-}
-
-.operator {
-  background-color: orange;
-}
-
-.operator:hover {
-  background-color: rgb(252, 195, 121);
-}
-
-.button-zero {
-  grid-column: 1/3;
-  height: 100%;
-}
+} */
