+++ src/App.js
@@ -2,8 +2,7 @@
 
 import './App.css';
 import React, { Component } from 'react';
-import Navbar from './components/Navbar';
-// import Home from './components/Home';
+import Navbar from './components/js/Navbar';
 
 export default class App extends Component {
   render() {
