--- src/components/Home.js
@@ -1,16 +1,0 @@
-import NavLinks from './NavLinks';
-
-const Home = () => (
-  <>
-    <NavLinks />
-    <h2>Welcome to our page!</h2>
-    <p>
-      Lorem ipsum dolor sit amet, consectetur adipisicing elit. Voluptates blanditiis culpa
-      laborum in? Eligendi ea enim ipsum. Ducimus, doloribus? Aspernatur nulla quod, facere
-      voluptate esse explicabo est reprehenderit maxime temporibus harum cupiditate dolorem
-      id laborum a ea asperiores doloribus blanditiis quasi recusandae voluptas eveniet! Animi
-    </p>
-  </>
-);
-
-export default Home;
