+++ src/components/css/Header.css
@@ -1,0 +1,26 @@
+header {
+  display: flex;
+  justify-content: space-between;
+  background-color: #32234a;
+  padding: 10px;
+}
+
+.nav-item-color {
+  color: rgb(255, 98, 150);
+}
+
+.navbar-heading {
+  margin-left: 30px;
+}
+
+.nav-links {
+  display: flex;
+  margin: 0 30px;
+  align-items: center;
+}
+
+.nav-item {
+  margin-right: 40px;
+  text-decoration: none;
+  font-weight: 700;
+}
