+++ src/components/css/Home.css
@@ -1,0 +1,7 @@
+.homepage-content {
+  padding: 0 20px;
+}
+
+.homepage-heading {
+  margin: 40px 0;
+}
