+++ src/components/js/Home.js
@@ -1,0 +1,27 @@
+import '../css/Home.css';
+import NavLinks from './NavLinks';
+
+const Home = () => (
+  <>
+    <NavLinks />
+    <section className="homepage-content">
+      <h2 className="homepage-heading">Welcome to our page!</h2>
+      <p className="homepage-paragraph">
+        They are among the simpler interactive software tools, and, as such,
+        they provide operations for the user to select one at a time. They
+        can be used to perform any process that consists of a sequence of
+        steps each of which applies one of these operations, and have no
+        purpose other than these processes, because the operations are the
+        sole, or at least the primary, features of the calculator, rather
+        than being secondary features that support other functionality that
+        is not normally known simply as calculation.
+        As a calculator, rather than a computer, they
+        usually have a small set of relatively simple
+        operations, perform short processes that are not compute intensive
+        and do not accept large amounts of input data or produce many results.
+      </p>
+    </section>
+  </>
+);
+
+export default Home;
